refactor: in-place organization fixes from the review's structure pass
Placement fixes within the existing single assembly, staged ahead of
the FanaBridge.Core split:

- Col03FrameClassifier moves Protocol/ -> Transport/: it is the reader
  thread's router predicate set, and its old placement made Transport
  and Protocol mutually dependent (the folders read as layers but
  weren't). Namespace intentionally unchanged (FanaBridge.Protocol) —
  this reorg renames no namespaces.
- DiagnosticsReport + ConverterCaptureProbe move to a new Diagnostics/
  folder: the report spans HID inventory (HidSharp), DirectInput
  (SharpDX), identity decode, and markdown formatting — naming the
  folder makes that sanctioned exception explicit instead of letting
  Protocol/ mean two things.
- SettingsControl.EncodeText moves to SevenSegment.EncodeWithDots
  (verbatim fold logic; resolves the self-flagged TODO): the display
  test's scroll path now shares the protocol layer's folding instead
  of maintaining a private copy. DisplayEncoder.DisplayText unchanged.
- Profiles/LedDefinition.cs no longer hides InputMapping and
  LedChannelConverter — each significant type gets its own file
  (LedChannelConverter has its own test file naming a type that
  previously couldn't be found by filename).
- Test fakes gather under TestDoubles/: FakeReportStream, and the
  Control Mapper fakes incl. the load-bearing type-name shim
  (SimHub.Plugins.OutputPlugins.ControlRemapper.ControlMapperPlugin,
  resolved by full name via reflection) which was buried at the tail
  of ControlMapperBridgeTests.cs.

Adds four SevenSegment.EncodeWithDots tests (466 total).
