+++ src/FanaBridge/Diagnostics/ConverterCaptureProbe.cs
