+++ src/FanaBridge/Diagnostics/DiagnosticsReport.cs
