+++ src/FanaBridge/Profiles/LedDefinition.cs
@@ -1,46 +1,7 @@
-using System;
-using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace FanaBridge.Profiles
 {
-    /// <summary>
-    /// Structured input association for a single LED.
-    /// An encoder LED can carry BOTH relative and absolute mappings
-    /// simultaneously so the user can switch modes at runtime.
-    /// </summary>
-    public class InputMapping
-    {
-        /// <summary>
-        /// Button input ID (e.g. "JoystickPlugin.FANATEC_Wheel.Button3").
-        /// Populated only for momentary push-button LEDs.
-        /// </summary>
-        [JsonProperty("button", NullValueHandling = NullValueHandling.Ignore)]
-        public string Button { get; set; }
-
-        /// <summary>
-        /// Relative (incremental) encoder inputs: [CW, CCW].
-        /// Null/omitted if not captured or not applicable.
-        /// </summary>
-        [JsonProperty("relative", NullValueHandling = NullValueHandling.Ignore)]
-        public List<string> Relative { get; set; }
-
-        /// <summary>
-        /// Absolute (positional) encoder inputs — one entry per detent
-        /// in the order they were detected (typically 12 for Fanatec).
-        /// Null/omitted if not captured or not applicable.
-        /// </summary>
-        [JsonProperty("absolute", NullValueHandling = NullValueHandling.Ignore)]
-        public List<string> Absolute { get; set; }
-
-        /// <summary>True when at least one input has been captured.</summary>
-        [JsonIgnore]
-        public bool HasAny =>
-            Button != null ||
-            (Relative != null && Relative.Count > 0) ||
-            (Absolute != null && Absolute.Count > 0);
-    }
-
     /// <summary>
     /// Describes a single physical LED on the device.
     /// The array order in the profile defines the SimHub logical index.
@@ -55,7 +16,7 @@ namespace FanaBridge.Profiles
         /// <summary>
         /// Index within the channel's protocol array.
         /// For <see cref="LedChannel.ButtonRgb"/>: slot in the subcmd 0x02 color array (0-11).
-        /// For <see cref="LedChannel.ButtonAuxIntensity"/>: byte index in the 16-byte intensity payload.
+        /// For <see cref="LedChannel.ButtonAuxIntensity"/>: byte index in the intensity payload.
         /// For <see cref="LedChannel.RevRgb"/>/<see cref="LedChannel.FlagRgb"/>: slot in subcmd 0x00/0x01.
         /// </summary>
         [JsonProperty("hwIndex")]
@@ -84,42 +45,5 @@ namespace FanaBridge.Profiles
         /// </summary>
         [JsonProperty("inputMapping", NullValueHandling = NullValueHandling.Ignore)]
         public InputMapping InputMapping { get; set; }
-    }
-
-    /// <summary>
-    /// JSON converter for <see cref="LedChannel"/> that accepts both v1 and v2
-    /// channel names. V1 names (rev, flag, color, mono) are mapped to their
-    /// v2 equivalents during deserialization.
-    /// </summary>
-    internal class LedChannelConverter : JsonConverter<LedChannel>
-    {
-        public override LedChannel ReadJson(JsonReader reader, Type objectType, LedChannel existingValue, bool hasExistingValue, JsonSerializer serializer)
-        {
-            string value = reader.Value as string;
-            if (value == null)
-                throw new JsonSerializationException("LED channel must be a string.");
-
-            // Try v2 names first (standard enum parse, case-insensitive)
-            if (Enum.TryParse(value, true, out LedChannel channel))
-                return channel;
-
-            // Fall back to v1 name mapping (only names that shipped in v1)
-            switch (value.ToLowerInvariant())
-            {
-                case "rev": return LedChannel.RevRgb;
-                case "flag": return LedChannel.FlagRgb;
-                case "color": return LedChannel.ButtonRgb;
-                case "mono": return LedChannel.ButtonAuxIntensity;
-                default:
-                    throw new JsonSerializationException($"Unknown LED channel '{value}'.");
-            }
-        }
-
-        public override void WriteJson(JsonWriter writer, LedChannel value, JsonSerializer serializer)
-        {
-            // Always write v2 names in camelCase
-            string name = value.ToString();
-            writer.WriteValue(char.ToLowerInvariant(name[0]) + name.Substring(1));
-        }
     }
 }
