+++ tests/FanaBridge.Tests/TestDoubles/FakeReportStream.cs
